+++ api/lib/models.ts
@@ -14,6 +14,10 @@ export default class Models {
     ConnectionSink: Modeler<typeof pgtypes.ConnectionSink>;
     ConnectionToken: Modeler<typeof pgtypes.ConnectionToken>;
 
+    Setting: Modeler<typeof pgtypes.Setting>;
+
+    Overlay: Modeler<typeof pgtypes.Overlay>;
+
     Profile: Modeler<typeof pgtypes.Profile>;
     ProfileChat: ProfileChat;
     ProfileOverlay: Modeler<typeof pgtypes.ProfileOverlay>;
@@ -30,6 +34,8 @@ export default class Models {
         this.Icon = new Icon(pg);
 
         this.Token = new Modeler(pg, pgtypes.Token);
+        this.Setting = new Modeler(pg, pgtypes.Setting);
+        this.Overlay = new Modeler(pg, pgtypes.Overlay);
         this.Server = new Modeler(pg, pgtypes.Server);
         this.Profile = new Modeler(pg, pgtypes.Profile);
         this.ProfileOverlay = new Modeler(pg, pgtypes.ProfileOverlay);