+++ api/lib/schema.ts
@@ -178,6 +178,11 @@ export const LayerAlert = pgTable('layer_alerts', {
     hidden: boolean('hidden').notNull().default(false)
 });
 
+export const Setting = pgTable('settings', {
+    key: text('key').primaryKey(),
+    value: text('value').notNull().default('')
+});
+
 export const Server = pgTable('server', {
     id: serial('id').primaryKey(),
     created: timestamp('created', { withTimezone: true, mode: 'string' }).notNull().default(sql`Now()`),
@@ -219,6 +224,16 @@ export const ProfileMission = pgTable('profile_missions', {
     token: text('token').notNull(),
     created: timestamp('created', { withTimezone: true, mode: 'string' }).notNull().default(sql`Now()`),
     updated: timestamp('updated', { withTimezone: true, mode: 'string' }).notNull().default(sql`Now()`),
+});
+
+export const Overlay = pgTable('overlays', {
+    id: serial('id').primaryKey(),
+    name: text('name').notNull(),
+    created: timestamp('created', { withTimezone: true, mode: 'string' }).notNull().default(sql`Now()`),
+    updated: timestamp('updated', { withTimezone: true, mode: 'string' }).notNull().default(sql`Now()`),
+    type: text('type').notNull().default('vector'),
+    styles: json('styles'),
+    url: text('url').notNull()
 });
 
 export const ProfileOverlay = pgTable('profile_overlays', {