+++ api/routes/config.ts
@@ -1,0 +1,33 @@
+import { Type } from '@sinclair/typebox'
+import Schema from '@openaddresses/batch-schema';
+import Err from '@openaddresses/batch-error';
+import Auth from '../lib/auth.js';
+import Config from '../lib/config.js';
+
+export default async function router(schema: Schema, config: Config) {
+    await schema.get('/config', {
+        name: 'Get Config',
+        group: 'Config',
+        description: 'Get Config',
+        query: Type.Object({
+            keys: Type.String()
+        }),
+        res: Type.Any()
+    }, async (req, res) => {
+        try {
+            await Auth.as_user(config, req, { admin: true });
+
+            const final: Record<string, string> = {};
+            (await Promise.allSettled((req.query.keys.split(',').map((key) => {
+                return config.models.Setting.from(key);
+            })))).forEach((k) => {
+                if (k.status === 'rejected') return;
+                return final[k.value.key] = k.value.value;
+            });
+
+            return res.json(final);
+        } catch (err) {
+            return Err.respond(err, res);
+        }
+    });
+}